Crate axaddrspace

ArceOS-Hypervisor guest VM address space management module.

Structs

AddrSpace

The virtual memory address space.

GuestPhysAddr

Guest physical address.

GuestVirtAddr

Guest virtual address.

MappingFlags

Generic page table entry flags that indicate the corresponding mapped memory region permissions and attributes.

NestedPageFaultInfo

Information about nested page faults.

Enums

Backend

A unified enum type for different memory mapping backends.

Type Aliases

GuestPhysAddrRange

Guest physical address range.

GuestVirtAddrRange

Guest virtual address range.

HostPhysAddr

Host physical address.

HostVirtAddr

Host virtual address.