Crate axaddrspace

ArceOS-Hypervisor guest VM address space management module.

Modules

device

Definitions about device accessing.

Structs

AddrSpace

The virtual memory address space.

GuestPhysAddr

Guest physical address.

GuestVirtAddr

Guest virtual address.

MappingFlags

Generic page table entry flags that indicate the corresponding mapped memory region permissions and attributes.

NestedPageFaultInfo

Information about nested page faults.

PhysFrame

A physical frame which will be automatically deallocated when dropped.

Enums

Backend

A unified enum type for different memory mapping backends.

Traits

AxMmHal

Hardware abstraction layer for memory management.

GuestMemoryAccessor

A stateful accessor to the memory space of a guest

Type Aliases

GuestPhysAddrRange

Guest physical address range.

GuestVirtAddrRange

Guest virtual address range.

HostPhysAddr

Host physical address.

HostVirtAddr

Host virtual address.